requirements.txt
playwright>=1.48.0
anthropic>=0.40.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
